Fixed more relative imports and got rid of "long" calls.
Fixed a couple of relative imports to explicitly import, for example,
spanlib.util instead of just util or .util. Also got rid of a call of the form
"long(some_int)" as all integers in Python 3+ are long, and the long() method
no longer exists.

+++ lib/spanlib/data.py
@@ -32,7 +32,7 @@ except:
     has_cdat_support = False
 
 default_missing_value = npy.ma.default_fill_value(0.)
-from util import Logger, dict_filter, broadcast
+from spanlib.util import Logger, dict_filter, broadcast
 
 
 class Data(Logger):
@@ -177,7 +177,8 @@ class Data(Logger):
         count[count<minvalid] = 0 # <minvalid -> 0
         count = npy.clip(count, 0, 1)
         # - save as 0/1
-        self.ns = long(count.sum())
+        # self.ns = long(count.sum())
+        self.ns = count.sum()
         self.compress = count.size != self.ns
         self.good = count>0 # points in space where there are enough data in time
         self.minvalid = self.nvalid = minvalid
